+++ packages/fhi-designsystem/src/components/typography/fhi-display/fhi-display.test.ts
@@ -58,11 +58,10 @@ describe('fhi-display', () => {
       component = await fixture(html`
         <fhi-display level="2">Test</fhi-display>
       `);
-      expect(component).shadowDom.to.equal(`
-      <h2 class="display">
-        <slot></slot>
-      </h2>
-    `);
+
+      const heading = component.shadowRoot?.querySelector('h2');
+
+      expect(heading?.tagName).to.equal('H2');
     });
   });